+++ app/Support/System/SystemBackupArchivePackage.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Support\System;
+
+class SystemBackupArchivePackage
+{
+    public const PRODUCT = 'WebBlocks CMS';
+
+    public const PACKAGE_TYPE = 'system_backup';
+
+    public const FEATURE_VERSION = 1;
+
+    public const FORMAT_VERSION = 1;
+}